SQL-4133: Move some sql_parsers.rs parse functions into parsers
