+++ query/src/sql/parsers/parser_explain.rs
@@ -1,0 +1,49 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Borrow from apache/arrow/rust/datafusion/src/sql/sql_parser
+// See notice.md
+
+use common_planners::ExplainType;
+use sqlparser::parser::ParserError;
+use sqlparser::tokenizer::Token;
+
+use crate::sql::statements::DfExplain;
+use crate::sql::DfParser;
+use crate::sql::DfStatement;
+
+impl<'a> DfParser<'a> {
+    // Parse an SQL EXPLAIN statement.
+    pub(crate) fn parse_explain(&mut self) -> Result<DfStatement, ParserError> {
+        // Parser is at the token immediately after EXPLAIN
+        // Check for EXPLAIN VERBOSE
+        let typ = match self.parser.peek_token() {
+            Token::Word(w) => match w.value.to_uppercase().as_str() {
+                "PIPELINE" => {
+                    self.parser.next_token();
+                    ExplainType::Pipeline
+                }
+                "GRAPH" => {
+                    self.parser.next_token();
+                    ExplainType::Graph
+                }
+                _ => ExplainType::Syntax,
+            },
+            _ => ExplainType::Syntax,
+        };
+
+        let statement = Box::new(self.parse_query()?);
+        Ok(DfStatement::Explain(DfExplain { typ, statement }))
+    }
+}
