+++ query/src/sql/parsers/parser_insert.rs
@@ -1,0 +1,56 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Borrow from apache/arrow/rust/datafusion/src/sql/sql_parser
+// See notice.md
+
+use sqlparser::ast::Statement;
+use sqlparser::parser::ParserError;
+
+use crate::parser_err;
+use crate::sql::statements::DfInsertStatement;
+use crate::sql::DfParser;
+use crate::sql::DfStatement;
+
+impl<'a> DfParser<'a> {
+    pub(crate) fn parse_insert(&mut self) -> Result<DfStatement, ParserError> {
+        self.parser.next_token();
+        match self.parser.parse_insert()? {
+            Statement::Insert {
+                or,
+                table_name,
+                columns,
+                overwrite,
+                source,
+                partitioned,
+                format,
+                after_columns,
+                table,
+                on,
+            } => Ok(DfStatement::InsertQuery(DfInsertStatement {
+                or,
+                table_name,
+                columns,
+                overwrite,
+                source,
+                partitioned,
+                format,
+                after_columns,
+                table,
+                on,
+            })),
+            _ => parser_err!("Expect set insert statement"),
+        }
+    }
+}
