+++ query/src/sql/sql_parser.rs
@@ -16,82 +16,30 @@
 // See notice.md
 
 use std::collections::HashMap;
-use std::convert::TryFrom;
 use std::time::Instant;
 
 use common_exception::ErrorCode;
-use common_io::prelude::OptionsDeserializer;
-use common_meta_types::Credentials;
-use common_meta_types::FileFormat;
-use common_meta_types::StageParams;
-use common_meta_types::UserIdentity;
-use common_meta_types::UserPrivilegeSet;
-use common_meta_types::UserPrivilegeType;
-use common_planners::ExplainType;
-use common_planners::Optimization;
 use metrics::histogram;
-use serde::Deserialize;
-use sqlparser::ast::ColumnDef;
-use sqlparser::ast::ColumnOptionDef;
-use sqlparser::ast::Ident;
-use sqlparser::ast::Statement;
-use sqlparser::ast::TableConstraint;
 use sqlparser::ast::Value;
 use sqlparser::dialect::keywords::Keyword;
 use sqlparser::dialect::Dialect;
 use sqlparser::dialect::GenericDialect;
-use sqlparser::parser::IsOptional;
 use sqlparser::parser::Parser;
 use sqlparser::parser::ParserError;
 use sqlparser::tokenizer::Token;
 use sqlparser::tokenizer::Tokenizer;
 use sqlparser::tokenizer::Whitespace;
-use sqlparser::tokenizer::Word;
 
-use super::statements::DfCopy;
-use super::statements::DfDescribeStage;
-use crate::sql::statements::DfAlterUDF;
-use crate::sql::statements::DfAlterUser;
-use crate::sql::statements::DfAuthOption;
-use crate::sql::statements::DfCreateDatabase;
-use crate::sql::statements::DfCreateStage;
-use crate::sql::statements::DfCreateTable;
-use crate::sql::statements::DfCreateUDF;
-use crate::sql::statements::DfCreateUser;
-use crate::sql::statements::DfDescribeTable;
-use crate::sql::statements::DfDropDatabase;
-use crate::sql::statements::DfDropStage;
-use crate::sql::statements::DfDropTable;
-use crate::sql::statements::DfDropUDF;
-use crate::sql::statements::DfDropUser;
-use crate::sql::statements::DfExplain;
-use crate::sql::statements::DfGrantObject;
-use crate::sql::statements::DfGrantStatement;
-use crate::sql::statements::DfInsertStatement;
-use crate::sql::statements::DfKillStatement;
-use crate::sql::statements::DfOptimizeTable;
-use crate::sql::statements::DfQueryStatement;
-use crate::sql::statements::DfRevokeStatement;
-use crate::sql::statements::DfSetVariable;
-use crate::sql::statements::DfShowCreateDatabase;
-use crate::sql::statements::DfShowCreateTable;
-use crate::sql::statements::DfShowDatabases;
 use crate::sql::statements::DfShowEngines;
-use crate::sql::statements::DfShowFunctions;
-use crate::sql::statements::DfShowGrants;
-use crate::sql::statements::DfShowKind;
 use crate::sql::statements::DfShowMetrics;
 use crate::sql::statements::DfShowProcessList;
 use crate::sql::statements::DfShowSettings;
-use crate::sql::statements::DfShowTables;
 use crate::sql::statements::DfShowUsers;
-use crate::sql::statements::DfTruncateTable;
-use crate::sql::statements::DfUseDatabase;
-use crate::sql::statements::DfUseTenant;
 use crate::sql::DfHint;
 use crate::sql::DfStatement;
 
 // Use `Parser::expected` instead, if possible
+#[macro_export]
 macro_rules! parser_err {
     ($MSG:expr) => {
         Err(ParserError::ParserError($MSG.to_string().into()))
@@ -100,7 +48,7 @@ macro_rules! parser_err {
 
 /// SQL Parser
 pub struct DfParser<'a> {
-    parser: Parser<'a>,
+    pub(crate) parser: Parser<'a>,
 }
 
 impl<'a> DfParser<'a> {
@@ -173,7 +121,7 @@ impl<'a> DfParser<'a> {
     }
 
     /// Report unexpected token
-    fn expected<T>(&self, expected: &str, found: Token) -> Result<T, ParserError> {
+    pub(crate) fn expected<T>(&self, expected: &str, found: Token) -> Result<T, ParserError> {
         parser_err!(format!("Expected {}, found: {}", expected, found))
     }
 
@@ -253,7 +201,7 @@ impl<'a> DfParser<'a> {
                         "USE" => self.parse_use_database(),
                         "KILL" => self.parse_kill_query(),
                         "OPTIMIZE" => self.parse_optimize(),
-                        "SUDO" => self.parse_sudo_command(),
+                        "SUDO" => self.parse_admin_command(),
                         _ => self.expected("Keyword", self.parser.peek_token()),
                     },
                     _ => self.expected("an SQL statement", Token::Word(w)),
@@ -264,184 +212,24 @@ impl<'a> DfParser<'a> {
         }
     }
 
-    fn parse_query(&mut self) -> Result<DfStatement, ParserError> {
-        // self.parser.prev_token();
-        let native_query = self.parser.parse_query()?;
-        Ok(DfStatement::Query(Box::new(DfQueryStatement::try_from(
-            native_query,
-        )?)))
-    }
-
-    fn parse_set(&mut self) -> Result<DfStatement, ParserError> {
-        self.parser.next_token();
-        match self.parser.parse_set()? {
-            Statement::SetVariable {
-                local,
-                hivevar,
-                variable,
-                value,
-            } => Ok(DfStatement::SetVariable(DfSetVariable {
-                local,
-                hivevar,
-                variable,
-                value,
-            })),
-            _ => parser_err!("Expect set Variable statement"),
-        }
-    }
-
-    fn parse_insert(&mut self) -> Result<DfStatement, ParserError> {
-        self.parser.next_token();
-        match self.parser.parse_insert()? {
-            Statement::Insert {
-                or,
-                table_name,
-                columns,
-                overwrite,
-                source,
-                partitioned,
-                format,
-                after_columns,
-                table,
-                on,
-            } => Ok(DfStatement::InsertQuery(DfInsertStatement {
-                or,
-                table_name,
-                columns,
-                overwrite,
-                source,
-                partitioned,
-                format,
-                after_columns,
-                table,
-                on,
-            })),
-            _ => parser_err!("Expect set insert statement"),
-        }
-    }
-
-    /// Parse an SQL EXPLAIN statement.
-    pub fn parse_explain(&mut self) -> Result<DfStatement, ParserError> {
-        // Parser is at the token immediately after EXPLAIN
-        // Check for EXPLAIN VERBOSE
-        let typ = match self.parser.peek_token() {
-            Token::Word(w) => match w.value.to_uppercase().as_str() {
-                "PIPELINE" => {
-                    self.parser.next_token();
-                    ExplainType::Pipeline
+    fn parse_create(&mut self) -> Result<DfStatement, ParserError> {
+        match self.parser.next_token() {
+            Token::Word(w) => {
+                //TODO:make stage to sql parser keyword
+                if w.value.to_uppercase() == "STAGE" {
+                    self.parse_create_stage()
+                } else {
+                    match w.keyword {
+                        Keyword::TABLE => self.parse_create_table(),
+                        Keyword::DATABASE => self.parse_create_database(),
+                        Keyword::USER => self.parse_create_user(),
+                        Keyword::FUNCTION => self.parse_create_udf(),
+                        _ => self.expected("create statement", Token::Word(w)),
+                    }
                 }
-                "GRAPH" => {
-                    self.parser.next_token();
-                    ExplainType::Graph
-                }
-                _ => ExplainType::Syntax,
-            },
-            _ => ExplainType::Syntax,
-        };
-
-        let statement = Box::new(self.parse_query()?);
-        Ok(DfStatement::Explain(DfExplain { typ, statement }))
-    }
-
-    // parse show tables.
-    fn parse_show_tables(&mut self) -> Result<DfStatement, ParserError> {
-        let tok = self.parser.next_token();
-        match &tok {
-            Token::EOF | Token::SemiColon => Ok(DfStatement::ShowTables(DfShowTables::create(
-                DfShowKind::All,
-            ))),
-            Token::Word(w) => match w.keyword {
-                Keyword::LIKE => Ok(DfStatement::ShowTables(DfShowTables::create(
-                    DfShowKind::Like(self.parser.parse_identifier()?),
-                ))),
-                Keyword::WHERE => Ok(DfStatement::ShowTables(DfShowTables::create(
-                    DfShowKind::Where(self.parser.parse_expr()?),
-                ))),
-                Keyword::FROM | Keyword::IN => Ok(DfStatement::ShowTables(DfShowTables::create(
-                    DfShowKind::FromOrIn(self.parser.parse_object_name()?),
-                ))),
-                _ => self.expected("like or where", tok),
-            },
-            _ => self.expected("like or where", tok),
-        }
-    }
-
-    // parse show databases where database = xxx or where database
-    fn parse_show_databases(&mut self) -> Result<DfStatement, ParserError> {
-        let tok = self.parser.next_token();
-        match &tok {
-            Token::EOF | Token::SemiColon => Ok(DfStatement::ShowDatabases(
-                DfShowDatabases::create(DfShowKind::All),
-            )),
-            Token::Word(w) => match w.keyword {
-                Keyword::LIKE => Ok(DfStatement::ShowDatabases(DfShowDatabases::create(
-                    DfShowKind::Like(self.parser.parse_identifier()?),
-                ))),
-                Keyword::WHERE => Ok(DfStatement::ShowDatabases(DfShowDatabases::create(
-                    DfShowKind::Where(self.parser.parse_expr()?),
-                ))),
-                Keyword::FROM | Keyword::IN => Ok(DfStatement::ShowDatabases(
-                    DfShowDatabases::create(DfShowKind::FromOrIn(self.parser.parse_object_name()?)),
-                )),
-                _ => self.expected("like or where", tok),
-            },
-            _ => self.expected("like or where", tok),
-        }
-    }
-
-    // parse show functions statement
-    fn parse_show_functions(&mut self) -> Result<DfStatement, ParserError> {
-        let tok = self.parser.next_token();
-        match &tok {
-            Token::EOF | Token::SemiColon => Ok(DfStatement::ShowFunctions(
-                DfShowFunctions::create(DfShowKind::All),
-            )),
-            Token::Word(w) => match w.keyword {
-                Keyword::LIKE => Ok(DfStatement::ShowFunctions(DfShowFunctions::create(
-                    DfShowKind::Like(self.parser.parse_identifier()?),
-                ))),
-                Keyword::WHERE => Ok(DfStatement::ShowFunctions(DfShowFunctions::create(
-                    DfShowKind::Where(self.parser.parse_expr()?),
-                ))),
-                _ => self.expected("like or where", tok),
-            },
-            _ => self.expected("like or where", tok),
-        }
-    }
-
-    // This is a copy of the equivalent implementation in sqlparser.
-    fn parse_columns(&mut self) -> Result<(Vec<ColumnDef>, Vec<TableConstraint>), ParserError> {
-        let mut columns = vec![];
-        let mut constraints = vec![];
-        if !self.parser.consume_token(&Token::LParen) || self.parser.consume_token(&Token::RParen) {
-            return Ok((columns, constraints));
-        }
-
-        loop {
-            if let Some(constraint) = self.parser.parse_optional_table_constraint()? {
-                constraints.push(constraint);
-            } else if let Token::Word(_) = self.parser.peek_token() {
-                let column_def = self.parse_column_def()?;
-                columns.push(column_def);
-            } else {
-                return self.expected(
-                    "column name or constraint definition",
-                    self.parser.peek_token(),
-                );
             }
-            let comma = self.parser.consume_token(&Token::Comma);
-            if self.parser.consume_token(&Token::RParen) {
-                // allow a trailing comma, even though it's not in standard
-                break;
-            } else if !comma {
-                return self.expected(
-                    "',' or ')' after column definition",
-                    self.parser.peek_token(),
-                );
-            }
+            unexpected => self.expected("create statement", unexpected),
         }
-
-        Ok((columns, constraints))
     }
 
     /// This is a copy from sqlparser
@@ -493,60 +281,6 @@ impl<'a> DfParser<'a> {
         }
     }
 
-    fn parse_column_def(&mut self) -> Result<ColumnDef, ParserError> {
-        let name = self.parser.parse_identifier()?;
-        let data_type = self.parser.parse_data_type()?;
-        let collation = if self.parser.parse_keyword(Keyword::COLLATE) {
-            Some(self.parser.parse_object_name()?)
-        } else {
-            None
-        };
-        let mut options = vec![];
-        loop {
-            if self.parser.parse_keyword(Keyword::CONSTRAINT) {
-                let name = Some(self.parser.parse_identifier()?);
-                if let Some(option) = self.parser.parse_optional_column_option()? {
-                    options.push(ColumnOptionDef { name, option });
-                } else {
-                    return self.expected(
-                        "constraint details after CONSTRAINT <name>",
-                        self.parser.peek_token(),
-                    );
-                }
-            } else if let Some(option) = self.parser.parse_optional_column_option()? {
-                options.push(ColumnOptionDef { name: None, option });
-            } else {
-                break;
-            };
-        }
-        Ok(ColumnDef {
-            name,
-            data_type,
-            collation,
-            options,
-        })
-    }
-
-    fn parse_create(&mut self) -> Result<DfStatement, ParserError> {
-        match self.parser.next_token() {
-            Token::Word(w) => {
-                //TODO:make stage to sql parser keyword
-                if w.value.to_uppercase() == "STAGE" {
-                    self.parse_create_stage()
-                } else {
-                    match w.keyword {
-                        Keyword::TABLE => self.parse_create_table(),
-                        Keyword::DATABASE => self.parse_create_database(),
-                        Keyword::USER => self.parse_create_user(),
-                        Keyword::FUNCTION => self.parse_create_udf(),
-                        _ => self.expected("create statement", Token::Word(w)),
-                    }
-                }
-            }
-            unexpected => self.expected("create statement", unexpected),
-        }
-    }
-
     fn parse_alter(&mut self) -> Result<DfStatement, ParserError> {
         match self.parser.next_token() {
             Token::Word(w) => match w.keyword {
@@ -558,34 +292,13 @@ impl<'a> DfParser<'a> {
         }
     }
 
-    fn parse_create_database(&mut self) -> Result<DfStatement, ParserError> {
-        let if_not_exists =
-            self.parser
-                .parse_keywords(&[Keyword::IF, Keyword::NOT, Keyword::EXISTS]);
-        let name = self.parser.parse_object_name()?;
-        let (engine, engine_options) = self.parse_database_engine()?;
-
-        let create = DfCreateDatabase {
-            if_not_exists,
-            name,
-            engine,
-            engine_options,
-            options: HashMap::new(),
-        };
-
-        Ok(DfStatement::CreateDatabase(create))
-    }
-
     fn parse_describe(&mut self) -> Result<DfStatement, ParserError> {
         if self.consume_token("stage") {
-            let obj_name = self.parser.parse_object_name()?;
-            let desc = DfDescribeStage { name: obj_name };
-            Ok(DfStatement::DescribeStage(desc))
+            self.parse_desc_stage()
+        } else if self.consume_token("table") {
+            self.parse_desc_table()
         } else {
-            self.consume_token("table");
-            let table_name = self.parser.parse_object_name()?;
-            let desc = DfDescribeTable { name: table_name };
-            Ok(DfStatement::DescribeTable(desc))
+            self.expected("stage or table", self.parser.peek_token())
         }
     }
 
@@ -609,692 +322,29 @@ impl<'a> DfParser<'a> {
         }
     }
 
-    /// Drop database.
-    fn parse_drop_database(&mut self) -> Result<DfStatement, ParserError> {
-        let if_exists = self.parser.parse_keywords(&[Keyword::IF, Keyword::EXISTS]);
-        let db_name = self.parser.parse_object_name()?;
-
-        let drop = DfDropDatabase {
-            if_exists,
-            name: db_name,
-        };
-
-        Ok(DfStatement::DropDatabase(drop))
-    }
-
-    /// Drop table.
-    fn parse_drop_table(&mut self) -> Result<DfStatement, ParserError> {
-        let if_exists = self.parser.parse_keywords(&[Keyword::IF, Keyword::EXISTS]);
-        let table_name = self.parser.parse_object_name()?;
-
-        let drop = DfDropTable {
-            if_exists,
-            name: table_name,
-        };
-
-        Ok(DfStatement::DropTable(drop))
-    }
-
-    // Parse 'sudo ...'.
-    fn parse_sudo_command(&mut self) -> Result<DfStatement, ParserError> {
-        self.parser.next_token();
-        match self.consume_token("USE") {
-            true if self.consume_token("TENANT") => self.parse_use_tenant(),
-            _ => self.expected("Unsupported sudo command", self.parser.peek_token()),
-        }
-    }
-
-    // Parse 'sudo use tenant [tenant id]'.
-    fn parse_use_tenant(&mut self) -> Result<DfStatement, ParserError> {
-        let name = self.parser.parse_object_name()?;
-        Ok(DfStatement::UseTenant(DfUseTenant { name }))
-    }
-
-    // Parse 'use database' db name.
-    fn parse_use_database(&mut self) -> Result<DfStatement, ParserError> {
-        if !self.consume_token("USE") {
-            return self.expected("Must USE", self.parser.peek_token());
-        }
-        let name = self.parser.parse_object_name()?;
-        Ok(DfStatement::UseDatabase(DfUseDatabase { name }))
-    }
-
-    // Parse 'KILL statement'.
-    fn parse_kill<const KILL_QUERY: bool>(&mut self) -> Result<DfStatement, ParserError> {
-        Ok(DfStatement::KillStatement(DfKillStatement {
-            object_id: self.parser.parse_identifier()?,
-            kill_query: KILL_QUERY,
-        }))
-    }
-
-    // Parse 'KILL statement'.
-    fn parse_kill_query(&mut self) -> Result<DfStatement, ParserError> {
-        match self.consume_token("KILL") {
-            true if self.consume_token("QUERY") => self.parse_kill::<true>(),
-            true if self.consume_token("CONNECTION") => self.parse_kill::<false>(),
-            true => self.parse_kill::<false>(),
-            false => self.expected("Must KILL", self.parser.peek_token()),
-        }
-    }
-
-    fn parse_create_user(&mut self) -> Result<DfStatement, ParserError> {
-        let if_not_exists =
-            self.parser
-                .parse_keywords(&[Keyword::IF, Keyword::NOT, Keyword::EXISTS]);
-        let name = self.parser.parse_literal_string()?;
-        let hostname = if self.consume_token("@") {
-            self.parser.parse_literal_string()?
-        } else {
-            String::from("%")
-        };
-
-        let auth_option = self.parse_auth_option()?;
-
-        let create = DfCreateUser {
-            if_not_exists,
-            name,
-            hostname,
-            auth_options: auth_option,
-        };
-
-        Ok(DfStatement::CreateUser(create))
-    }
-
-    fn parse_alter_user(&mut self) -> Result<DfStatement, ParserError> {
-        let if_current_user = self.consume_token("USER")
-            && self.parser.expect_token(&Token::LParen).is_ok()
-            && self.parser.expect_token(&Token::RParen).is_ok();
-        let name = if !if_current_user {
-            self.parser.parse_literal_string()?
-        } else {
-            String::from("")
-        };
-        let hostname = if !if_current_user {
-            if self.consume_token("@") {
-                self.parser.parse_literal_string()?
-            } else {
-                String::from("%")
-            }
-        } else {
-            String::from("")
-        };
-
-        let auth_option = self.parse_auth_option()?;
-
-        let alter = DfAlterUser {
-            if_current_user,
-            name,
-            hostname,
-            auth_option,
-        };
-
-        Ok(DfStatement::AlterUser(alter))
-    }
-
-    fn parse_drop_user(&mut self) -> Result<DfStatement, ParserError> {
-        let if_exists = self.parser.parse_keywords(&[Keyword::IF, Keyword::EXISTS]);
-        let name = self.parser.parse_literal_string()?;
-        let hostname = if self.consume_token("@") {
-            self.parser.parse_literal_string()?
-        } else {
-            String::from("%")
-        };
-        let drop = DfDropUser {
-            if_exists,
-            name,
-            hostname,
-        };
-        Ok(DfStatement::DropUser(drop))
-    }
-
-    fn parse_auth_option(&mut self) -> Result<DfAuthOption, ParserError> {
-        let exist_not = self.parser.parse_keyword(Keyword::NOT);
-        let exist_identified = self.consume_token("IDENTIFIED");
-
-        if exist_not {
-            if !exist_identified {
-                parser_err!("expect IDENTIFIED after NOT")
-            } else {
-                Ok(DfAuthOption::no_password())
-            }
-        } else if !exist_identified {
-            Ok(DfAuthOption::default())
-        } else {
-            let arg_with = if self.consume_token("WITH") {
-                Some(self.parser.parse_literal_string()?)
-            } else {
-                None
-            };
-            if arg_with == Some("no_password".to_string()) {
-                Ok(DfAuthOption::no_password())
-            } else {
-                let auth_string = if self.parser.parse_keyword(Keyword::BY) {
-                    Some(self.parser.parse_literal_string()?)
-                } else {
-                    None
-                };
-                Ok(DfAuthOption {
-                    auth_type: arg_with,
-                    by_value: auth_string,
-                })
-            }
-        }
-    }
-
-    fn parse_stage_file_format(&mut self) -> Result<FileFormat, ParserError> {
-        let options = if self.consume_token("FILE_FORMAT") {
-            self.parser.expect_token(&Token::Eq)?;
-            self.parser.expect_token(&Token::LParen)?;
-            let options = self.parse_options()?;
-
-            self.parser.expect_token(&Token::RParen)?;
-
-            options
-        } else {
-            HashMap::new()
-        };
-        let file_format = FileFormat::deserialize(OptionsDeserializer::new(&options))
-            .map_err(|e| ParserError::ParserError(format!("Invalid file format options: {}", e)))?;
-        Ok(file_format)
-    }
-
-    fn parse_stage_credentials(&mut self) -> Result<Credentials, ParserError> {
-        let options = if self.consume_token("CREDENTIALS") {
-            self.parser.expect_token(&Token::Eq)?;
-            self.parser.expect_token(&Token::LParen)?;
-            let options = self.parse_options()?;
-            self.parser.expect_token(&Token::RParen)?;
-
-            options
-        } else {
-            HashMap::new()
-        };
-
-        let credentials = Credentials::deserialize(OptionsDeserializer::new(&options))
-            .map_err(|e| ParserError::ParserError(format!("Invalid credentials options: {}", e)))?;
-        Ok(credentials)
-    }
-
-    fn parse_create_stage(&mut self) -> Result<DfStatement, ParserError> {
-        let if_not_exists =
-            self.parser
-                .parse_keywords(&[Keyword::IF, Keyword::NOT, Keyword::EXISTS]);
-        let name = self.parser.parse_literal_string()?;
-        let url = if self.consume_token("URL") {
-            self.parser.expect_token(&Token::Eq)?;
-            self.parser.parse_literal_string()?
-        } else {
-            return parser_err!("Missing URL");
-        };
-
-        if !url.to_uppercase().starts_with("S3") {
-            return parser_err!("Not supported storage");
-        }
-
-        let credentials = self.parse_stage_credentials()?;
-        let stage_params = StageParams::new(url.as_str(), credentials);
-        let file_format = self.parse_stage_file_format()?;
-
-        let comments = if self.consume_token("COMMENTS") {
-            self.parser.expect_token(&Token::Eq)?;
-            self.parser.parse_literal_string()?
-        } else {
-            String::from("")
-        };
-
-        let create = DfCreateStage {
-            if_not_exists,
-            stage_name: name,
-            stage_params,
-            file_format,
-            comments,
-        };
-
-        Ok(DfStatement::CreateStage(create))
-    }
-
-    fn parse_drop_stage(&mut self) -> Result<DfStatement, ParserError> {
-        let if_exists = self.parser.parse_keywords(&[Keyword::IF, Keyword::EXISTS]);
-        let stage_name = self.parser.parse_literal_string()?;
-
-        let drop = DfDropStage {
-            if_exists,
-            stage_name,
-        };
-        Ok(DfStatement::DropStage(drop))
-    }
-
-    fn parse_udf_parameters(&mut self) -> Result<Vec<String>, ParserError> {
-        let mut params = vec![];
-        let mut found_right_paren = false;
-        let mut expect_next_param = false;
-
-        self.parser.expect_token(&Token::LParen)?;
-        loop {
-            let next_token = self.parser.next_token();
-            if next_token == Token::EOF || next_token == Token::SemiColon {
-                break;
-            }
-            if next_token == Token::RParen {
-                found_right_paren = true;
-                break;
-            }
-
-            match next_token {
-                Token::Word(Word {
-                    value,
-                    quote_style,
-                    keyword,
-                }) => {
-                    if keyword != Keyword::NoKeyword {
-                        return parser_err!(format!(
-                            "Keyword can not be parameter, got: {}",
-                            value
-                        ));
-                    }
-                    if let Some(quote) = quote_style {
-                        return parser_err!(format!(
-                            "Quote is not allowed in parameters, remove: {}",
-                            quote
-                        ));
-                    }
-                    if params.contains(&value) {
-                        return parser_err!(format!(
-                            "Duplicate parameter is not allowed, keep only one: {}",
-                            &value
-                        ));
-                    }
-
-                    expect_next_param = false;
-                    params.push(value);
-                }
-                Token::Comma => {
-                    expect_next_param = true;
-                }
-                other => {
-                    return parser_err!(format!("Expect words or comma, but got: {}", other));
-                }
-            }
-        }
-
-        if !found_right_paren {
-            return parser_err!("Can not find complete parameters, `)` is missing");
-        }
-        if expect_next_param {
-            return parser_err!("Found a redundant `,` in the parameters");
-        }
-
-        Ok(params)
-    }
-
-    fn parse_udf_definition_expr(&mut self, until_token: Vec<&str>) -> Result<String, ParserError> {
-        // Match ->
-        self.parser.expect_token(&Token::Minus)?;
-        let next_token = self.parser.next_token_no_skip();
-        if next_token != Some(&Token::Gt) {
-            return parser_err!(format!("Expected >, found: {:#?}", next_token));
-        }
-
-        let definition = self.consume_token_until_or_end(until_token).join("");
-        if definition.is_empty() {
-            return parser_err!("UDF definition can not be empty");
-        }
-
-        Ok(definition)
-    }
-
-    fn parse_udf_desc(&mut self, desc_token: &str) -> Result<String, ParserError> {
-        if self.consume_token(desc_token) {
-            self.parser.expect_token(&Token::Eq)?;
-            Ok(self.parser.parse_literal_string()?)
-        } else {
-            Ok(String::from(""))
-        }
-    }
-
-    fn parse_create_udf(&mut self) -> Result<DfStatement, ParserError> {
-        let if_not_exists =
-            self.parser
-                .parse_keywords(&[Keyword::IF, Keyword::NOT, Keyword::EXISTS]);
-
-        let udf_name = self.parser.parse_literal_string()?;
-        self.parser.expect_keyword(Keyword::AS)?;
-
-        let desc_token = "DESC";
-        let parameters = self.parse_udf_parameters()?;
-        let definition = self.parse_udf_definition_expr(vec![desc_token])?;
-
-        let description = self.parse_udf_desc(desc_token)?;
-        let create_udf = DfCreateUDF {
-            if_not_exists,
-            udf_name,
-            parameters,
-            definition,
-            description,
-        };
-
-        Ok(DfStatement::CreateUDF(create_udf))
-    }
-
-    fn parse_alter_udf(&mut self) -> Result<DfStatement, ParserError> {
-        let udf_name = self.parser.parse_literal_string()?;
-        let as_token = Token::make_keyword("AS");
-        self.parser.expect_token(&as_token)?;
-
-        let desc_token = "DESC";
-        let parameters = self.parse_udf_parameters()?;
-        let definition = self.parse_udf_definition_expr(vec![desc_token])?;
-
-        let description = self.parse_udf_desc(desc_token)?;
-        let update_udf = DfAlterUDF {
-            udf_name,
-            parameters,
-            definition,
-            description,
-        };
-
-        Ok(DfStatement::AlterUDF(update_udf))
-    }
-
-    fn parse_drop_udf(&mut self) -> Result<DfStatement, ParserError> {
-        let if_exists = self.parser.parse_keywords(&[Keyword::IF, Keyword::EXISTS]);
-        let udf_name = self.parser.parse_literal_string()?;
-
-        let drop_udf = DfDropUDF {
-            if_exists,
-            udf_name,
-        };
-        Ok(DfStatement::DropUDF(drop_udf))
-    }
-
-    fn parse_create_table(&mut self) -> Result<DfStatement, ParserError> {
-        let if_not_exists =
-            self.parser
-                .parse_keywords(&[Keyword::IF, Keyword::NOT, Keyword::EXISTS]);
-        let table_name = self.parser.parse_object_name()?;
-
-        // Parse the table which we copy schema from. This is for create table like statement.
-        // https://dev.mysql.com/doc/refman/8.0/en/create-table-like.html
-        let mut table_like = None;
-        if self.parser.parse_keyword(Keyword::LIKE) {
-            table_like = Some(self.parser.parse_object_name()?);
-        }
-
-        let (columns, _) = self.parse_columns()?;
-        if !columns.is_empty() && table_like.is_some() {
-            return parser_err!("mix create table like statement and column definition.");
-        }
-
-        let engine = self.parse_table_engine()?;
-
-        // parse table options: https://dev.mysql.com/doc/refman/8.0/en/create-table.html
-        let options = self.parse_options()?;
-
-        let mut query = None;
-        if let Token::Word(Word { keyword, .. }) = self.parser.peek_token() {
-            let mut has_query = false;
-            if keyword == Keyword::AS {
-                self.parser.next_token();
-                has_query = true;
-            }
-            if has_query || keyword == Keyword::SELECT {
-                let native = self.parser.parse_query()?;
-                query = Some(Box::new(DfQueryStatement::try_from(native)?))
-            }
-        }
-
-        let create = DfCreateTable {
-            if_not_exists,
-            name: table_name,
-            columns,
-            engine,
-            options,
-            like: table_like,
-            query,
-        };
-
-        Ok(DfStatement::CreateTable(create))
-    }
-
-    fn parse_database_engine(&mut self) -> Result<(String, HashMap<String, String>), ParserError> {
-        // TODO make ENGINE as a keyword
-        if !self.consume_token("ENGINE") {
-            return Ok(("".to_string(), HashMap::new()));
-        }
-
-        self.parser.expect_token(&Token::Eq)?;
-        let engine = self.parser.next_token().to_string();
-        let options = if self.parser.consume_token(&Token::LParen) {
-            let options = self.parse_options()?;
-            self.parser.expect_token(&Token::RParen)?;
-            options
-        } else {
-            HashMap::new()
-        };
-        Ok((engine, options))
-    }
-
-    /// Parses the set of valid formats
-    fn parse_table_engine(&mut self) -> Result<String, ParserError> {
-        // TODO make ENGINE as a keyword
-        if !self.consume_token("ENGINE") {
-            return Ok("FUSE".to_string());
-        }
-
-        self.parser.expect_token(&Token::Eq)?;
-        Ok(self.parser.next_token().to_string())
-    }
-
     fn parse_show_create(&mut self) -> Result<DfStatement, ParserError> {
         match self.parser.next_token() {
             Token::Word(w) => match w.keyword {
-                Keyword::TABLE => {
-                    let table_name = self.parser.parse_object_name()?;
-
-                    let show_create_table = DfShowCreateTable { name: table_name };
-                    Ok(DfStatement::ShowCreateTable(show_create_table))
-                }
-                Keyword::DATABASE => {
-                    let db_name = self.parser.parse_object_name()?;
-
-                    let show_create_database = DfShowCreateDatabase { name: db_name };
-                    Ok(DfStatement::ShowCreateDatabase(show_create_database))
-                }
+                Keyword::TABLE => self.parse_show_create_table(),
+                Keyword::DATABASE => self.parse_show_create_database(),
                 _ => self.expected("show create statement", Token::Word(w)),
             },
             unexpected => self.expected("show create statement", unexpected),
         }
     }
 
-    fn parse_show_grants(&mut self) -> Result<DfStatement, ParserError> {
-        // SHOW GRANTS
-        if !self.consume_token("FOR") {
-            return Ok(DfStatement::ShowGrants(DfShowGrants {
-                user_identity: None,
-            }));
-        }
-
-        // SHOW GRANTS FOR 'u1'@'%'
-        let (username, hostname) = self.parse_user_identity()?;
-        Ok(DfStatement::ShowGrants(DfShowGrants {
-            user_identity: Some(UserIdentity { username, hostname }),
-        }))
-    }
-
     fn parse_truncate(&mut self) -> Result<DfStatement, ParserError> {
         self.parser.next_token();
         match self.parser.next_token() {
             Token::Word(w) => match w.keyword {
-                Keyword::TABLE => {
-                    let table_name = self.parser.parse_object_name()?;
-                    let purge = self.parser.parse_keyword(Keyword::PURGE);
-                    let statement = DfTruncateTable {
-                        name: table_name,
-                        purge,
-                    };
-                    Ok(DfStatement::TruncateTable(statement))
-                }
+                Keyword::TABLE => self.parse_truncate_table(),
                 _ => self.expected("truncate statement", Token::Word(w)),
             },
             unexpected => self.expected("truncate statement", unexpected),
         }
     }
 
-    fn parse_privileges(&mut self) -> Result<UserPrivilegeSet, ParserError> {
-        let mut privileges = UserPrivilegeSet::empty();
-        loop {
-            match self.parser.next_token() {
-                Token::Word(w) => match w.keyword {
-                    // Keyword::USAGE => privileges.set_privilege(UserPrivilegeType::Usage),
-                    Keyword::CREATE => {
-                        if self.consume_token("USER") {
-                            privileges.set_privilege(UserPrivilegeType::CreateUser)
-                        } else if self.consume_token("ROLE") {
-                            privileges.set_privilege(UserPrivilegeType::CreateRole)
-                        } else {
-                            privileges.set_privilege(UserPrivilegeType::Create)
-                        }
-                    }
-                    Keyword::DROP => privileges.set_privilege(UserPrivilegeType::Drop),
-                    Keyword::ALTER => privileges.set_privilege(UserPrivilegeType::Alter),
-                    Keyword::SELECT => privileges.set_privilege(UserPrivilegeType::Select),
-                    Keyword::INSERT => privileges.set_privilege(UserPrivilegeType::Insert),
-                    Keyword::UPDATE => privileges.set_privilege(UserPrivilegeType::Update),
-                    Keyword::DELETE => privileges.set_privilege(UserPrivilegeType::Delete),
-                    // TODO: uncomment this after sqlparser-rs accepts the SUPER keyword
-                    // Keyword::SUPER => privileges.set_privilege(UserPrivilegeType::Super)
-                    Keyword::GRANT => privileges.set_privilege(UserPrivilegeType::Grant),
-                    Keyword::ALL => {
-                        privileges.set_all_privileges();
-                        // GRANT ALL [PRIVILEGES]
-                        self.consume_token("PRIVILEGES");
-                        break;
-                    }
-                    _ => return self.expected("privilege type", Token::Word(w)),
-                },
-                unexpected => return self.expected("privilege type", unexpected),
-            };
-            if !self.parser.consume_token(&Token::Comma) {
-                break;
-            }
-        }
-        Ok(privileges)
-    }
-
-    fn parse_grant(&mut self) -> Result<DfStatement, ParserError> {
-        let privileges = self.parse_privileges()?;
-        if !self.parser.parse_keyword(Keyword::ON) {
-            return self.expected("keyword ON", self.parser.peek_token());
-        }
-        let on = self.parse_grant_object()?;
-        if !self.parser.parse_keyword(Keyword::TO) {
-            return self.expected("keyword TO", self.parser.peek_token());
-        }
-        let (name, hostname) = self.parse_user_identity()?;
-        let grant = DfGrantStatement {
-            name,
-            hostname,
-            on,
-            priv_types: privileges,
-        };
-        Ok(DfStatement::GrantPrivilege(grant))
-    }
-
-    fn parse_revoke(&mut self) -> Result<DfStatement, ParserError> {
-        let privileges = self.parse_privileges()?;
-        if !self.parser.parse_keyword(Keyword::ON) {
-            return self.expected("keyword ON", self.parser.peek_token());
-        }
-        let on = self.parse_grant_object()?;
-        if !self.parser.parse_keyword(Keyword::FROM) {
-            return self.expected("keyword FROM", self.parser.peek_token());
-        }
-        let (username, hostname) = self.parse_user_identity()?;
-        let revoke = DfRevokeStatement {
-            username,
-            hostname,
-            on,
-            priv_types: privileges,
-        };
-        Ok(DfStatement::RevokePrivilege(revoke))
-    }
-
-    fn parse_user_identity(&mut self) -> Result<(String, String), ParserError> {
-        let username = self.parser.parse_literal_string()?;
-        let hostname = if self.consume_token("@") {
-            self.parser.parse_literal_string()?
-        } else {
-            String::from("%")
-        };
-        Ok((username, hostname))
-    }
-
-    /// Parse a possibly qualified, possibly quoted identifier or wild card, e.g.
-    /// `*` or `myschema`.*. The sub string pattern like "db0%" is not in planned.
-    fn parse_grant_object(&mut self) -> Result<DfGrantObject, ParserError> {
-        let chunk0 = self.parse_grant_object_pattern_chunk()?;
-        // "*" as current db or "table" with current db
-        if !self.consume_token(".") {
-            if chunk0.value == "*" {
-                return Ok(DfGrantObject::Database(None));
-            }
-            return Ok(DfGrantObject::Table(None, chunk0.value));
-        }
-        let chunk1 = self.parse_grant_object_pattern_chunk()?;
-
-        // *.* means global
-        if chunk1.value == "*" && chunk0.value == "*" {
-            return Ok(DfGrantObject::Global);
-        }
-        // *.table is not allowed
-        if chunk0.value == "*" {
-            return self.expected("whitespace", Token::Period);
-        }
-        // db.*
-        if chunk1.value == "*" {
-            return Ok(DfGrantObject::Database(Some(chunk0.value)));
-        }
-        // db.table
-        Ok(DfGrantObject::Table(Some(chunk0.value), chunk1.value))
-    }
-
-    /// Parse a chunk from the object pattern, it might be * or an identifier
-    pub fn parse_grant_object_pattern_chunk(&mut self) -> Result<Ident, ParserError> {
-        if self.consume_token("*") {
-            return Ok(Ident::new("*"));
-        }
-        let token = self.parser.peek_token();
-        self.parser
-            .parse_identifier()
-            .or_else(|_| self.expected("identifier or *", token))
-    }
-
-    // copy into mycsvtable
-    // from @my_ext_stage/tutorials/dataloading/contacts1.csv format CSV [options];
-    fn parse_copy(&mut self) -> Result<DfStatement, ParserError> {
-        self.parser.expect_keyword(Keyword::INTO)?;
-        let name = self.parser.parse_object_name()?;
-        let columns = self
-            .parser
-            .parse_parenthesized_column_list(IsOptional::Optional)?;
-        self.parser.expect_keyword(Keyword::FROM)?;
-        let location = self.parser.parse_literal_string()?;
-
-        self.parser.expect_keyword(Keyword::FORMAT)?;
-        let format = self.parser.next_token().to_string();
-
-        let options = self.parse_options()?;
-
-        Ok(DfStatement::Copy(DfCopy {
-            name,
-            columns,
-            location,
-            format,
-            options,
-        }))
-    }
-
-    fn parse_options(&mut self) -> Result<HashMap<String, String>, ParserError> {
+    pub(crate) fn parse_options(&mut self) -> Result<HashMap<String, String>, ParserError> {
         let mut options = HashMap::new();
         loop {
             let name = self.parser.parse_identifier();
@@ -1315,31 +365,7 @@ impl<'a> DfParser<'a> {
         Ok(options)
     }
 
-    fn parse_optimize(&mut self) -> Result<DfStatement, ParserError> {
-        // syntax: "optimize TABLE t [purge | compact | all]",  default action is "purge"
-        self.expect_token("OPTIMIZE")?;
-        self.parser.expect_keyword(Keyword::TABLE)?;
-        let object_name = self.parser.parse_object_name()?;
-        let operation = match self.parser.next_token() {
-            Token::EOF => Ok(Optimization::PURGE),
-            Token::Word(w) => match w.keyword {
-                Keyword::ALL => Ok(Optimization::ALL),
-                Keyword::PURGE => Ok(Optimization::PURGE),
-                Keyword::NoKeyword if w.value.to_uppercase().as_str() == "COMPACT" => {
-                    Ok(Optimization::COMPACT)
-                }
-                _ => self.expected("one of PURGE, COMPACT, ALL", Token::Word(w)),
-            },
-            t => self.expected("Nothing, or one of PURGE, COMPACT, ALL", t),
-        }?;
-
-        Ok(DfStatement::OptimizeTable(DfOptimizeTable {
-            name: object_name,
-            operation,
-        }))
-    }
-
-    fn consume_token(&mut self, expected: &str) -> bool {
+    pub(crate) fn consume_token(&mut self, expected: &str) -> bool {
         if self.parser.peek_token().to_string().to_uppercase() == *expected.to_uppercase() {
             self.parser.next_token();
             true
@@ -1348,7 +374,7 @@ impl<'a> DfParser<'a> {
         }
     }
 
-    fn consume_token_until_or_end(&mut self, until_tokens: Vec<&str>) -> Vec<String> {
+    pub(crate) fn consume_token_until_or_end(&mut self, until_tokens: Vec<&str>) -> Vec<String> {
         let mut tokens = vec![];
 
         loop {
@@ -1367,7 +393,7 @@ impl<'a> DfParser<'a> {
         tokens
     }
 
-    fn expect_token(&mut self, expected: &str) -> Result<(), ParserError> {
+    pub(crate) fn expect_token(&mut self, expected: &str) -> Result<(), ParserError> {
         if self.consume_token(expected) {
             Ok(())
         } else {
